[not verified] Licensing: Move licensing endpoints to package
